requirements.txt
h5py==3.9.0
ipykernel==6.25.2
ipython==8.15.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
keras==2.13.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
nbformat==5.9.2
numpy==1.24.3
pandas==2.1.0
pickleshare==0.7.5
Pillow==10.0.0
scikit-learn==1.3.0
seaborn==0.12.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0